Moved base search functionality into separate class
